Add support for conditional rendering of video areas
With the new system for selecting the layout separately for each cut
depending on which type of content is present, it is also possible to
make the layout selection depend on other things relating to the current
state of the meeting.

The EDL and layout data structures are updated to support "conditions".
When the EDL is generated, various conditions can be included which
identify things about the state of the session at that time. Then, when
selecting a layout, the layout is only used if the conditions specified
in the layout match the conditions in the session.

Two conditions are added initially:

* `presentation_used`: Indicates whether any actions have been done on
  the presentation/whiteboard area.
* `screenshare_as_content`: Indicates the current state of the
  presentation/screenshare swap feature.

The video recording format's code is simplified since it no longer needs
to detect whether the presentation area was used itself; it can rely on
the conditions to handle this instead. The default configuration file
has been updated with layouts which handle the two new conditions.