fix(livekit): broken volume control for plugin SDK and away mode
When using LiveKit as the audio bridge, output volume control is broken
for both the plugin SDK primitive and away mode (when set to
deafen-on-away).

Add proper output volume control in LK by properly reacting to speaker
level changes. This includes the intruction of a new storage variable
(speakerLevel) and a useSpeakerLevel hook so that LK's RoomAudioRender
can change its volume when necessary.