fix(livekit): SIP calls aren't accepted w/o published audio tracks
When livekit/sip dispatch rules do not require a PIN for inbound SIP
calls, they're only fully established once at least one audio
subscription in a room succeeds. In our use case, where PIN prompting
and IVR is currently done by an external gateway, this is not desirable
as it may lead to a deafened-muted scenario when calling into a meeting
with all-muted users.

Add a patch to livekit/sip that introduces a new configuration option:
`no_pin_wait_for_subscribe`. In our case, it's set to false - so calls
will be established regardless of whether there's published audio tracks
in a room.

The goal is to later make this into a dispatch rule option and submit it
to upstream LiveKit.