fix(livekit): stricter measures to prevent publish-unpublish loops
Our LK audio bridge is prone to publish-unpublish loops since some of
its primitives (setSenderTrackEnabled, liveChangeInputDevice,
updateAudioConstraints) are not idempotent - while audio-manager thinks
they are (because they should be). A sample of the issues this causes is
what commit b3e04278 mitigated.

Add a set of stricter measures to the LK audio bridge to nudge it closer
to actual idempotency:
  - Add publish skip checks based on whether the state outcome should be
    the same based on: input stream IDs, pending publish requests not
    pre-empted by the publish queue
  - Add publish generation counters so that subsequent requests do no
    trample over previous ones
  - Add a debounced-unpublish approach: tracks are now only unpublished
    Xs (configurable, default 5s) after the track has been muted. This
    serves a few goals:
      - Reducing unpublish noise on frequent mute/unmute cycles
      - Reducing the chance a unpublish-publish loop may happen
      - Reducing overhead for frequent mute/unmutes for an end user
        (same logic applied to transparent listen only hold/unhold
        debounces)

Those changes improve the situation significantly - even when testing
with a revert of b3e04278. It is not ideal, though. Our LK audio bridge
is an astoundigly bad piece of code which, in the future, should be
rewritten outside of the audio-manager framework using LK-provided
primitives only. We'll have more flexibility to do a simpler thing in
that scenario.