fix: LiveKit token update fails on database
When a new LK token is generated for an existing session, its insertion
in the database is failing due to the usage of forceInsert. This is not
the expected behavior - the token should be updated.
While the behavior is mostly benign as token re-generation does not yet
play a significant role, this might bork unrelated batched DB ops and
should be fixed.

Fix the issue by using insertOrUpdate so that LK tokens are properly
updated.