fix(livekit): trigger forced room reconnection on fatal errors
There are known scenarios that currently can only be "addressed" via
full LiveKit room reconnections. One of those is a race condition when
publishing tracks - mainly audio - that causes all subsequent publish
attempts to time out. While this is a media server issue, it should be
mitigated on our side until fixed.

A few mitigations have already been added to the client for the above
scenario (timeouts). While they've reduced the error occurrence to
0.5%-0.7%~ of total sessions, it can still happen. A fail-safe that
forces a reconnect is necessary for the remaining cases.

Add a mechanism that can trigger LiveKit room reconnects on demand via
custom browser events. When a reconnect is desired, the breakpoint
should fire a 'livekitFatalError'-named event with the error and a
source identifier as payload.
This is generic enough so that we can add or remove scenarios to it as
needed. The first use case is the above time out edge case.

This forced reconnect can be enabled via settings.yml's
`public.media.livekit.reconnectOnFatalFailures`. This is FALSE by
default for now while it undergoes further testing.